Move module level functions into datasource
