+++ packages/apollos-data-connector-rock/__mocks__/@apollosproject/config.js
@@ -22,7 +22,7 @@ ApolloServer.loadJs({
   },
   ROCK: {
     TIMEZONE: 'America/New_York',
-  }
+  },
 });
 
 export default ApolloServer;
